+++ json_serializable/test/test_sources/test_sources.dart
@@ -21,11 +21,13 @@ class ConfigurationImplicitDefaults {
   dateTimeUtc: false,
   disallowUnrecognizedKeys: false,
   explicitToJson: false,
-  fieldRename: FieldRename.none,
+  fieldRename: RenameType.none,
   ignoreUnannotated: false,
   includeIfNull: true,
   genericArgumentFactories: false,
   createJsonSchema: false,
+  unionDiscriminator: 'type',
+  unionRename: RenameType.none,
 )
 class ConfigurationExplicitDefaults {
   int? field;